+++ apps/enmeshed/lib/profiles/device/device.dart
@@ -1,5 +1,5 @@
 export 'device_detail/device_detail_screen.dart';
 export 'device_detail/widgets/device_detail_widgets.dart';
 export 'devices_overview/devices_screen.dart';
-export 'devices_overview/widgets/create_device.dart';
+export 'modals/create_device.dart';
 export 'widgets/device_widgets.dart';
