+++ apps/enmeshed/lib/profiles/device/modals/add_or_connect_device.dart
@@ -4,6 +4,12 @@ import 'package:enmeshed_runtime_bridge/enmeshed_runtime_bridge.dart';
 import 'package:enmeshed_types/enmeshed_types.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
+import 'package:go_router/go_router.dart';
+
+import '/core/core.dart';
+import 'create_device.dart';
+import 'device_onboarding.dart';
+import 'device_onboarding_safety_note.dart';
 
 Future<void> addDevice({required BuildContext context, required String accountId, required Future<void> Function() reload}) async {
   await _AddOrConnectDeviceModal.show(context: context, accountId: accountId, reload: reload, existingDeviceAndToken: null);
@@ -42,6 +48,7 @@ class _AddOrConnectDeviceModal extends StatefulWidget {
   }) async {
     await showModalBottomSheet<void>(
       context: context,
+      isScrollControlled: true,
       builder:
           (context) => ConstrainedBox(
             constraints: BoxConstraints(maxHeight: MediaQuery.sizeOf(context).height * 0.9),
@@ -55,8 +62,56 @@ class _AddOrConnectDeviceModal extends StatefulWidget {
 }
 
 class _AddOrConnectDeviceModalState extends State<_AddOrConnectDeviceModal> {
+  (DeviceDTO, TokenDTO)? deviceAndToken;
+  bool safetyShown = false;
+
+  @override
+  void initState() {
+    super.initState();
+
+    deviceAndToken = widget.existingDeviceAndToken;
+  }
+
   @override
   Widget build(BuildContext context) {
-    return const Placeholder();
+    return AnimatedSwitcher(
+      layoutBuilder: (Widget? currentChild, List<Widget> previousChildren) {
+        return AnimatedSize(
+          duration: const Duration(milliseconds: 300),
+          child: Stack(alignment: Alignment.center, children: [...previousChildren, if (currentChild != null) currentChild]),
+        );
+      },
+      duration: const Duration(milliseconds: 300),
+      reverseDuration: Duration.zero,
+      transitionBuilder: (child, animation) {
+        return SlideTransition(
+          position: animation.drive(Tween(begin: const Offset(1, 0), end: Offset.zero).chain(CurveTween(curve: Curves.easeInOut))),
+          child: child,
+        );
+      },
+      child: switch ((deviceAndToken, safetyShown)) {
+        (null, _) => CreateDevice(
+          accountId: widget.accountId,
+          setDeviceAndToken: (device, token) async {
+            await widget.reload();
+            setState(() => deviceAndToken = (device, token));
+          },
+        ),
+        (_, false) => DeviceOnboardingSafetyNote(goToNextPage: () => setState(() => safetyShown = true)),
+        ((final DeviceDTO device, final TokenDTO token), true) => DeviceOnboarding(
+          token: token,
+          deviceId: device.id,
+          accountReference: widget.accountId,
+          onDeviceOnboarded: () async {
+            await widget.reload();
+
+            if (!context.mounted) return;
+
+            showSuccessSnackbar(context: context, text: context.l10n.devices_otherDeviceOnboardedSuccess);
+            context.pop();
+          },
+        ),
+      },
+    );
   }
 }
