+++ apps/enmeshed/lib/profiles/device/widgets/device_widgets.dart
@@ -1,6 +1,2 @@
-export 'add_or_connect_device.dart';
 export 'device_card.dart';
-export 'device_onboarding.dart';
-export 'device_onboarding_safety_note.dart';
-export 'device_onboarding_success.dart';
 export 'device_status_bar.dart';
